Drain microtask queue
...
Okay I used setTimeout.
We were basically wondering why my subscription somehow breaks all other breakpointobservers. In the end, it seems that setTimeout would somehow fix this. I still don't know what it is, and if you have a better suggestion it's well appreciated. But this works, and isn't particularly slow. It's just a little ugly.